# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MealFlow API
  description: India-focused family meal planning app API
  version: 0.9.0
  license:
    name: MIT
servers:
  - url: http://localhost:3458
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mf_sid
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        display_name:
          type: string
security:
  - cookieAuth: []
paths:
  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                display_name:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/change-password:
    post:
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...